feat(Bank-sdk): Add Skonto edge case to proceed button event
PP-435
