+++ bank-sdk/sdk/src/main/java/net/gini/android/bank/sdk/capture/digitalinvoice/DigitalInvoiceFragment.kt
@@ -354,7 +354,6 @@ internal open class DigitalInvoiceFragment : Fragment(), DigitalInvoiceScreenCon
     override fun payButtonClicked() {
         tryShowAttachDocToTransactionDialog {
             presenter?.pay()
-            trackProceedTapped()
             trackSdkClosedEvent()
         }
     }
@@ -677,12 +676,6 @@ internal open class DigitalInvoiceFragment : Fragment(), DigitalInvoiceScreenCon
         )
     }
 
-    private fun trackProceedTapped() = runCatching {
-        userAnalyticsEventTracker.trackEvent(
-            UserAnalyticsEvent.PROCEED_TAPPED,
-            setOf(UserAnalyticsEventProperty.Screen(screenName))
-        )
-    }
 
     private fun trackSdkClosedEvent() = runCatching {
         userAnalyticsEventTracker.trackEvent(
